backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-jose==3.3.0
lxml==5.3.0
